+++ src/wtime.hpp
@@ -1,3 +1,3 @@
 #pragma once
 // current time in seconds
-float wtime(void);
+double wtime(void);
